Introduccion de la logica de la casa y los prefabs
Hay que fijarse todavía los prefabs ,hitboxes y como va a ser el encuentro en la casa( a discutir)

+++ Assets/Scripts/Bloque.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Bloque : MonoBehaviour
 {
-    public int size;
+    public float size;
     // Start is called before the first frame update
     void Start()
     {
